requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
arrow==1.4.0
binaryornot==0.4.4
certifi==2025.11.12
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.1
cookiecutter==2.6.0
exceptiongroup==1.3.0
fastapi==0.121.3
flet==0.28.3
flet-cli==0.28.3
flet-desktop==0.28.3
flet-web==0.28.3
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
oauthlib==3.3.1
packaging==25.0
pydantic==2.12.4
pydantic_core==2.41.5
Pygments==2.19.2
pypng==0.20220715.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.2.1
python-slugify==8.0.4
PyYAML==6.0.3
qrcode==7.4.2
repath==0.9.0
requests==2.32.5
rich==14.2.0
six==1.17.0
sniffio==1.3.1
starlette==0.50.0
text-unidecode==1.3
toml==0.10.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.38.0
uvloop==0.22.1
watchdog==4.0.2
watchfiles==1.1.1
websockets==15.0.1
